Update Robot.java
Saturday I found this to not work with the bot. Setting inverted motors was not functioning (lines 44, 45) so I set it to a comment. Line 77 setting the motor safety to true caused the drive motors to not move. The speeds for the winch was not set properly. Changed the value from 0.3 to 0.5 to balance the speeds between directions. 
                                                                                                          -Jacob Moriarity